pyproject.toml
[project]
name = "fi-claude"
version = "0.1.0"
description = "Legible, functional fixed-income pricers for BRL Pre-CDI swaps, inflation-linked bonds, cross-currency basis swaps, and TBAs"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"

dependencies = [
    "pydantic>=2.0",
    "python-dateutil>=2.9",
]

[project.optional-dependencies]
# QuantLib as optional utility — not architectural dependency
quantlib = ["QuantLib>=1.34"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.4",
]
all = ["fi-claude[quantlib,dev]"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fi_claude"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]
